+++ tool/src/org/antlr/v4/automata/ATNOptimizer.java
@@ -113,9 +113,9 @@ public class ATNOptimizer {
 					continue;
 				}
 
-				ATNState blockEndState = decision.transition(interval.a).target.transition(0).target;
+				ATNState blockEndState = decision.transition((int)interval.a).target.transition(0).target;
 				IntervalSet matchSet = new IntervalSet();
-				for (int j = interval.a; j <= interval.b; j++) {
+				for (int j = (int)interval.a; j <= interval.b; j++) {
 					Transition matchTransition = decision.transition(j).target.transition(0);
 					if (matchTransition instanceof NotSetTransition) {
 						throw new UnsupportedOperationException("Not yet implemented.");
@@ -127,18 +127,18 @@ public class ATNOptimizer {
 				Transition newTransition;
 				if (matchSet.getIntervals().size() == 1) {
 					if (matchSet.size() == 1) {
-						newTransition = new AtomTransition(blockEndState, matchSet.getMinElement());
+						newTransition = new AtomTransition(blockEndState, (int)matchSet.getMinElement());
 					} else {
 						Interval matchInterval = matchSet.getIntervals().get(0);
-						newTransition = new RangeTransition(blockEndState, matchInterval.a, matchInterval.b);
+						newTransition = new RangeTransition(blockEndState, (int)matchInterval.a, (int)matchInterval.b);
 					}
 				} else {
 					newTransition = new SetTransition(blockEndState, matchSet);
 				}
 
-				decision.transition(interval.a).target.setTransition(0, newTransition);
-				for (int j = interval.a + 1; j <= interval.b; j++) {
-					Transition removed = decision.removeTransition(interval.a + 1);
+				decision.transition((int)interval.a).target.setTransition(0, newTransition);
+				for (int j = (int)interval.a + 1; j <= interval.b; j++) {
+					Transition removed = decision.removeTransition((int)interval.a + 1);
 					atn.removeState(removed.target);
 					removedStates++;
 				}
