+++ tool/src/org/antlr/v4/codegen/model/SerializedATN.java
@@ -34,6 +34,7 @@ import org.antlr.v4.automata.ATNSerializer;
 import org.antlr.v4.codegen.OutputModelFactory;
 import org.antlr.v4.runtime.atn.ATN;
 import org.antlr.v4.runtime.misc.IntegerList;
+import org.antlr.v4.runtime.misc.LongList;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -45,8 +46,8 @@ public class SerializedATN extends OutputModelObject {
 		super(factory);
 		IntegerList data = ATNSerializer.getSerialized(factory.getGrammar(), atn);
 		serialized = new ArrayList<String>(data.size());
-		for (int c : data.toArray()) {
-			String encoded = factory.getGenerator().getTarget().encodeIntAsCharEscape(c == -1 ? Character.MAX_VALUE : c);
+		for (long c : data.toArray()) {
+			String encoded = factory.getGenerator().getTarget().encodeIntAsCharEscape(c == -1 ? Character.MAX_VALUE : (int)c);
 			serialized.add(encoded);
 		}
 //		System.out.println(ATNSerializer.getDecoded(factory.getGrammar(), atn));
